[refactor] custom hook 제작 game #775
