+++ components/game/GameResult.tsx
@@ -1,8 +1,7 @@
-import { useRouter } from 'next/router';
-import { useEffect, useState } from 'react';
-import { QueryClient, QueryClientProvider } from 'react-query';
 import { SeasonMode } from 'types/mainType';
+import { useGameResult } from 'hooks/game/useGameResult';
 import GameResultList from 'components/game/GameResultList';
+import { QueryClient, QueryClientProvider } from 'react-query';
 
 interface GameResultProps {
   mode?: SeasonMode;
@@ -10,13 +9,10 @@ interface GameResultProps {
 }
 
 export default function GameResult({ mode, season }: GameResultProps) {
-  const [path, setPath] = useState('');
   const queryClient = new QueryClient();
-  const router = useRouter();
-  const asPath = router.asPath;
-  const intraId = router.query.intraId;
+  const { path } = useGameResult({ mode, season });
 
-  const makeQuery = () => {
+  /* const makeQuery = () => {
     if (asPath === '/' || asPath.includes('token')) {
       return { count: 3, status: 'live', gameId: '' };
     }
@@ -31,27 +27,7 @@ export default function GameResult({ mode, season }: GameResultProps) {
       query.count = 5;
     }
     return query;
-  };
-
-  const makePath = () => {
-    if (asPath === '/' || asPath.includes('token')) {
-      setPath(`/pingpong/games?count=${3}&status=${'live'}&gameId=`);
-      return;
-    }
-    const userQuery = intraId ? `/users/${intraId}` : '';
-    const seasonQuery = mode === 'rank' && `season=${season}`;
-    const modeQuery = mode !== 'both' && `mode=${mode}`;
-    const countQuery = router.pathname === '/users/detail' && `count=${5}`;
-    const query = [modeQuery, seasonQuery, countQuery, 'gameId=']
-      .filter((item) => item)
-      .join('&');
-    setPath(`/pingpong/games${userQuery}?${query}`);
-    return;
-  };
-
-  useEffect(() => {
-    makePath();
-  }, [asPath, intraId, mode, season]);
+  }; */
 
   return (
     <div>
